enforcement/Cursor/.cursor__archived_2025-04-12/scripts/requirements.txt
watchdog>=3.0.0
PyYAML>=6.0













